+++ 10-05-2017/p6.pas
@@ -1,5 +1,5 @@
 VAR
-    C : INTEGER;
+    C : REAL;
 BEGIN
     readln(C);
     writelb('S: ', 4*pi*C*C);
